requirements.txt
django
djangorestframework
djangorestframework-simplejwt>=5.2.0
psycopg2